requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jobspy==1.1.65
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
